pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "social_media_aggregator"
version = "0.1.5"

[tool.setuptools]
packages = { find = { where = ["src"] } }

[tool.poetry]
name = "social_media_aggregator"
version = "0.1.5"
description = "This is a project that aggregates several social media accounts into a single space"
authors = ["José Viegas <jviegas6@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
pytest-cov = "^5.0.0"
# Add your development dependencies here